requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
groq>=0.5.0
sqlalchemy>=1.4.0
python-dotenv>=0.19.0
httpx>=0.24.0
tqdm
nltk
scikit-learn
plotly>=5.0.0
openpyxl>=3.0.0